* merge back changes from VST SDK 3.5:
		* fix some warnings with some compilers
		* add CAnimationSplashScreen class
		* animations can have an optional notification object for receiving a message when the animation has finished

+++ trunk/vstgui/lib/cdatabrowser.cpp
@@ -893,7 +893,7 @@ int32_t GenericStringListDataBrowserSource::dbGetNumRows (CDataBrowser* browser)
 {
 	if (dataBrowser == 0)
 		dataBrowser = browser;
-	return stringList->size ();
+	return (int32_t)stringList->size ();
 }
 
 //-----------------------------------------------------------------------------
